feat: use RTIC for main board
