--- boards/communication/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "communication"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-cortex-m = { version = "^0.7.0", features = ["critical-section-single-core"] }
-cortex-m-rt = "^0.7.0"
-cortex-m-rtic = "1.1.3"
-panic-halt = "0.2.0"
-fugit = "0.3.3"
-systick-monotonic = "1.0.1"
-defmt-rtt = "0.4.0"
-defmt = "0.3.2"
-postcard = "1.0.2"
-heapless = "0.7.16"
-common-arm = { path = "../../libraries/common-arm" }
-atsamd-hal = { workspace = true }
-messages = { path = "../../libraries/messages" }
