+++ boards/main/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "main_board"
+name = "main"
 version = "0.1.0"
 edition = "2021"
 
@@ -8,10 +8,14 @@ edition = "2021"
 [dependencies]
 cortex-m = { version = "^0.7.0", features = ["critical-section-single-core"] }
 cortex-m-rt = "^0.7.0"
+cortex-m-rtic = "1.1.3"
 panic-halt = "0.2.0"
-embedded-sdmmc = "0.3.0"
-embedded-hal = "0.2.7"
+fugit = "0.3.3"
+systick-monotonic = "1.0.1"
 defmt-rtt = "0.4.0"
 defmt = "0.3.2"
-atsamd-hal = { workspace = true }
+postcard = "1.0.2"
+heapless = "0.7.16"
 common-arm = { path = "../../libraries/common-arm" }
+atsamd-hal = { workspace = true }
+messages = { path = "../../libraries/messages" }
